penumbra: generate interfaces without the chain messages
