+++ crates/proto/src/gen/penumbra.core.component.governance.v1alpha1.serde.rs
@@ -236,7 +236,7 @@ impl serde::Serialize for ChangedAppParameters {
     {
         use serde::ser::SerializeStruct;
         let mut len = 0;
-        if self.chain_params.is_some() {
+        if self.sct_params.is_some() {
             len += 1;
         }
         if self.community_pool_params.is_some() {
@@ -261,8 +261,8 @@ impl serde::Serialize for ChangedAppParameters {
             len += 1;
         }
         let mut struct_ser = serializer.serialize_struct("penumbra.core.component.governance.v1alpha1.ChangedAppParameters", len)?;
-        if let Some(v) = self.chain_params.as_ref() {
-            struct_ser.serialize_field("chainParams", v)?;
+        if let Some(v) = self.sct_params.as_ref() {
+            struct_ser.serialize_field("sctParams", v)?;
         }
         if let Some(v) = self.community_pool_params.as_ref() {
             struct_ser.serialize_field("communityPoolParams", v)?;
@@ -295,8 +295,8 @@ impl<'de> serde::Deserialize<'de> for ChangedAppParameters {
         D: serde::Deserializer<'de>,
     {
         const FIELDS: &[&str] = &[
-            "chain_params",
-            "chainParams",
+            "sct_params",
+            "sctParams",
             "community_pool_params",
             "communityPoolParams",
             "governance_params",
@@ -315,7 +315,7 @@ impl<'de> serde::Deserialize<'de> for ChangedAppParameters {
 
         #[allow(clippy::enum_variant_names)]
         enum GeneratedField {
-            ChainParams,
+            SctParams,
             CommunityPoolParams,
             GovernanceParams,
             IbcParams,
@@ -345,7 +345,7 @@ impl<'de> serde::Deserialize<'de> for ChangedAppParameters {
                         E: serde::de::Error,
                     {
                         match value {
-                            "chainParams" | "chain_params" => Ok(GeneratedField::ChainParams),
+                            "sctParams" | "sct_params" => Ok(GeneratedField::SctParams),
                             "communityPoolParams" | "community_pool_params" => Ok(GeneratedField::CommunityPoolParams),
                             "governanceParams" | "governance_params" => Ok(GeneratedField::GovernanceParams),
                             "ibcParams" | "ibc_params" => Ok(GeneratedField::IbcParams),
@@ -372,7 +372,7 @@ impl<'de> serde::Deserialize<'de> for ChangedAppParameters {
                 where
                     V: serde::de::MapAccess<'de>,
             {
-                let mut chain_params__ = None;
+                let mut sct_params__ = None;
                 let mut community_pool_params__ = None;
                 let mut governance_params__ = None;
                 let mut ibc_params__ = None;
@@ -382,11 +382,11 @@ impl<'de> serde::Deserialize<'de> for ChangedAppParameters {
                 let mut funding_params__ = None;
                 while let Some(k) = map_.next_key()? {
                     match k {
-                        GeneratedField::ChainParams => {
-                            if chain_params__.is_some() {
-                                return Err(serde::de::Error::duplicate_field("chainParams"));
+                        GeneratedField::SctParams => {
+                            if sct_params__.is_some() {
+                                return Err(serde::de::Error::duplicate_field("sctParams"));
                             }
-                            chain_params__ = map_.next_value()?;
+                            sct_params__ = map_.next_value()?;
                         }
                         GeneratedField::CommunityPoolParams => {
                             if community_pool_params__.is_some() {
@@ -436,7 +436,7 @@ impl<'de> serde::Deserialize<'de> for ChangedAppParameters {
                     }
                 }
                 Ok(ChangedAppParameters {
-                    chain_params: chain_params__,
+                    sct_params: sct_params__,
                     community_pool_params: community_pool_params__,
                     governance_params: governance_params__,
                     ibc_params: ibc_params__,
@@ -5571,6 +5571,124 @@ impl<'de> serde::Deserialize<'de> for ProposalWithdraw {
             }
         }
         deserializer.deserialize_struct("penumbra.core.component.governance.v1alpha1.ProposalWithdraw", FIELDS, GeneratedVisitor)
+    }
+}
+impl serde::Serialize for Ratio {
+    #[allow(deprecated)]
+    fn serialize<S>(&self, serializer: S) -> std::result::Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        use serde::ser::SerializeStruct;
+        let mut len = 0;
+        if self.numerator != 0 {
+            len += 1;
+        }
+        if self.denominator != 0 {
+            len += 1;
+        }
+        let mut struct_ser = serializer.serialize_struct("penumbra.core.component.governance.v1alpha1.Ratio", len)?;
+        if self.numerator != 0 {
+            #[allow(clippy::needless_borrow)]
+            struct_ser.serialize_field("numerator", ToString::to_string(&self.numerator).as_str())?;
+        }
+        if self.denominator != 0 {
+            #[allow(clippy::needless_borrow)]
+            struct_ser.serialize_field("denominator", ToString::to_string(&self.denominator).as_str())?;
+        }
+        struct_ser.end()
+    }
+}
+impl<'de> serde::Deserialize<'de> for Ratio {
+    #[allow(deprecated)]
+    fn deserialize<D>(deserializer: D) -> std::result::Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        const FIELDS: &[&str] = &[
+            "numerator",
+            "denominator",
+        ];
+
+        #[allow(clippy::enum_variant_names)]
+        enum GeneratedField {
+            Numerator,
+            Denominator,
+            __SkipField__,
+        }
+        impl<'de> serde::Deserialize<'de> for GeneratedField {
+            fn deserialize<D>(deserializer: D) -> std::result::Result<GeneratedField, D::Error>
+            where
+                D: serde::Deserializer<'de>,
+            {
+                struct GeneratedVisitor;
+
+                impl<'de> serde::de::Visitor<'de> for GeneratedVisitor {
+                    type Value = GeneratedField;
+
+                    fn expecting(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+                        write!(formatter, "expected one of: {:?}", &FIELDS)
+                    }
+
+                    #[allow(unused_variables)]
+                    fn visit_str<E>(self, value: &str) -> std::result::Result<GeneratedField, E>
+                    where
+                        E: serde::de::Error,
+                    {
+                        match value {
+                            "numerator" => Ok(GeneratedField::Numerator),
+                            "denominator" => Ok(GeneratedField::Denominator),
+                            _ => Ok(GeneratedField::__SkipField__),
+                        }
+                    }
+                }
+                deserializer.deserialize_identifier(GeneratedVisitor)
+            }
+        }
+        struct GeneratedVisitor;
+        impl<'de> serde::de::Visitor<'de> for GeneratedVisitor {
+            type Value = Ratio;
+
+            fn expecting(&self, formatter: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+                formatter.write_str("struct penumbra.core.component.governance.v1alpha1.Ratio")
+            }
+
+            fn visit_map<V>(self, mut map_: V) -> std::result::Result<Ratio, V::Error>
+                where
+                    V: serde::de::MapAccess<'de>,
+            {
+                let mut numerator__ = None;
+                let mut denominator__ = None;
+                while let Some(k) = map_.next_key()? {
+                    match k {
+                        GeneratedField::Numerator => {
+                            if numerator__.is_some() {
+                                return Err(serde::de::Error::duplicate_field("numerator"));
+                            }
+                            numerator__ = 
+                                Some(map_.next_value::<::pbjson::private::NumberDeserialize<_>>()?.0)
+                            ;
+                        }
+                        GeneratedField::Denominator => {
+                            if denominator__.is_some() {
+                                return Err(serde::de::Error::duplicate_field("denominator"));
+                            }
+                            denominator__ = 
+                                Some(map_.next_value::<::pbjson::private::NumberDeserialize<_>>()?.0)
+                            ;
+                        }
+                        GeneratedField::__SkipField__ => {
+                            let _ = map_.next_value::<serde::de::IgnoredAny>()?;
+                        }
+                    }
+                }
+                Ok(Ratio {
+                    numerator: numerator__.unwrap_or_default(),
+                    denominator: denominator__.unwrap_or_default(),
+                })
+            }
+        }
+        deserializer.deserialize_struct("penumbra.core.component.governance.v1alpha1.Ratio", FIELDS, GeneratedVisitor)
     }
 }
 impl serde::Serialize for Tally {
